+++ packages/lavas-core-vue/test/fixtures/simple/lavas.config.js
@@ -35,17 +35,9 @@ module.exports = {
         }
     },
     serviceWorker: {
+        enable: true,
         swSrc: path.join(__dirname, 'core/service-worker.js'),
         swDest: path.join(BUILD_PATH, 'service-worker.js'),
-        globDirectory: BUILD_PATH,
-        globPatterns: [
-            '**/*.{html,js,css,eot,svg,ttf,woff}'
-        ],
-        globIgnores: [
-            'sw-register.js',
-            '**/*.map'
-        ],
-        appshellUrl: '/appshell',
-        dontCacheBustUrlsMatching: /\.\w{8}\./
+        appshellUrl: '/appshell'
     }
 };